chatbot/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pdfplumber==0.11.6
sentence-transformers==3.4.1
numpy==1.26.4
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.2
